+++ Source/Geometry/PointCloud.cpp
@@ -9,7 +9,7 @@ PointCloud::PointCloud()
 {
 }
 
-PointCloud::PointCloud(int size, float max_x, float max_y)
+PointCloud::PointCloud(int size, double max_x, double max_y)
 {
     for(int i = 0; i < size; i++)
     {
